Big change
New DB schema broke my app, refactored to meet new spec.

Controllers need to be cut down and more services are needed.

Customer info cut, customer info is taken from the logged in user.

+++ Project_Noah_Ritcey/src/main/java/com/example/project_noah_ritcey/controllers/ToppingController.java
@@ -8,6 +8,8 @@ import org.springframework.web.bind.annotation.*;
 
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
+
 
 @Controller
 @RequiredArgsConstructor
@@ -18,7 +20,7 @@ public class ToppingController {
 
 
     @PostMapping("/addTopping")
-    public String addTopping(@RequestParam String name, @RequestParam Float price, Model model) {
+    public String addTopping(@RequestParam String name, @RequestParam BigDecimal price, Model model) {
         try {
             toppingService.addTopping(name, price);
             model.addAttribute("success", "Topping added");
@@ -57,7 +59,7 @@ public class ToppingController {
     }
 
     @PostMapping("/editTopping/{id}")
-    public String editTopping(@PathVariable Integer id, @RequestParam String name, @RequestParam Float price, Model model) {
+    public String editTopping(@PathVariable Integer id, @RequestParam String name, @RequestParam BigDecimal price, Model model) {
         try{
             toppingService.editTopping(id, name, price);
             model.addAttribute("success", "Topping edited");
